% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/addSummaryColumns.R
\name{addSummaryColumns}
\alias{addSummaryColumns}
\title{Add water use summaries.}
\usage{
addSummaryColumns(awudsData)
}
\arguments{
\item{awudsDAta}{Dataframe returned (from \link{getAWUDSdump})}
}
\value{
AWUDS data with publicSupply, domestic, irrigation, thermoelectricPower, livestockAndAquaculture, industrial, and mining added
}
\description{
Adds water use summary columns to the awudsData dataframe.
}
\examples{
\dontrun{
 awudsData<-getAWUDSdump('http://nwis.usgs.gov/awuds/dump/countydata')
 awudsData<-addSummaryColumns(awudsData)
}
}

