% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getAWUDSdump.R
\name{getAWUDSdump}
\alias{getAWUDSdump}
\title{Load AWUDS data into a dataframe}
\usage{
getAWUDSdump(dataLocation, convertToNumerical = TRUE)
}
\arguments{
\item{dataLocation}{A local or network path to the data file.}

\item{convertToNumerical}{(optional) TRUE (default) or FALSE to try to convert non numerical
types (typically factors) to numerical.}
}
\value{
Data frame of AWUDS data.
}
\description{
Given an AWUDS CSV dump file, returns a dataframe containing all
collumns of data optionally converted to numerical data types.
}
\examples{
\dontrun{
 awudsData<-getAWUDSdump('http://nwis.usgs.gov/awuds/dump/countydata')
}
}

