% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getCountyLayer.R
\name{getCountyLayer}
\alias{getCountyLayer}
\title{Downloads and returns loaded county spatial layer for plotting}
\usage{
getCountyLayer()
}
\value{
A sp Spatial Data Frame with all US countiesloaded
}
\description{
Downloads and returns loaded county spatial layer for plotting
}

