% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summaryHistogram.R
\name{summaryHistogram}
\alias{summaryHistogram}
\title{Categoriy Summary Historgram}
\usage{
summaryHistogram(awudsData, summaryCategory, popNorm = FALSE,
  truncateLower = 0, binwidth = NULL)
}
\arguments{
\item{popNorm}{(optional) TRUE of FALSE (default) normalizes to population.}

\item{truncateLower}{(optional) (0 default) Truncates the data as it will be plotted (after normalization) to some lower bound.}

\item{binwidth}{(optional) (range/30 default) Passed directly to plotting function to set binwidth.}

\item{awudsDAta}{Dataframe returned (from \link{getAWUDSdump})}

\item{summaryCategroy}{A category to summarize. Only publicSupply has been implemented so far.}
}
\value{
ggplot histogram
}
\description{
Generates a historgram of a summary category for demonstration purposes.
}
\examples{
\dontrun{
 summaryHistogram(awudsData, 'publicSupply', popNorm=FALSE, truncateLower=50, binwidth=150)
 summaryHistogram(awudsData, 'publicSupply', popNorm=TRUE, truncateLower=1)
}
}

